# implementation/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenants/{tenantId}": {
      "get": {
        "operationId": "TenantController_get",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a tenant",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}/organisations": {
      "post": {
        "operationId": "TenantController_createOrganisation",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register an Organisation",
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantListingController_organisations",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List organisations",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/relying-parties": {
      "post": {
        "operationId": "TenantController_createRelyingParty",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a Relying Party (TEST registration data)",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}/rp-services": {
      "post": {
        "operationId": "TenantController_createService",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a Relying Party Service",
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantListingController_relyingPartyServices",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List Relying Party Services",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/rp-services/{serviceId}": {
      "get": {
        "operationId": "TenantController_getService",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a Relying Party Service",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}/rp-services/{serviceId}/intended-uses": {
      "post": {
        "operationId": "TenantController_createIntendedUse",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register an intended use",
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantListingController_intendedUses",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List the intended uses of one Relying Party Service",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/rp-services/{serviceId}/registration-certificates": {
      "post": {
        "operationId": "TenantController_recordRegistrationCertificate",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Record a registration certificate for one intended use",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}/rp-services/{serviceId}/instance": {
      "post": {
        "operationId": "TenantController_provisionInstance",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Provision the Relying Party Instance and import its access certificate",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}/presentation-policies": {
      "post": {
        "operationId": "PolicyController_create",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a presentation policy",
        "tags": [
          "Presentation policies"
        ]
      },
      "get": {
        "operationId": "TenantListingController_presentationPolicies",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List presentation policies",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/presentation-policies/{policyId}": {
      "get": {
        "operationId": "PolicyController_get",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a presentation policy and its versions",
        "tags": [
          "Presentation policies"
        ]
      }
    },
    "/v1/tenants/{tenantId}/presentation-policies/{policyId}/versions": {
      "post": {
        "operationId": "PolicyController_createVersion",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a presentation policy version",
        "tags": [
          "Presentation policies"
        ]
      }
    },
    "/v1/tenants/{tenantId}/presentation-policies/{policyId}/versions/{version}/publish": {
      "post": {
        "operationId": "PolicyController_publish",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Publish a draft policy version",
        "tags": [
          "Presentation policies"
        ]
      }
    },
    "/v1/presentations": {
      "post": {
        "operationId": "PresentationController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a presentation transaction",
        "tags": [
          "Presentations"
        ]
      },
      "get": {
        "operationId": "TransactionListingController_presentations",
        "parameters": [
          {
            "name": "policyId",
            "required": false,
            "in": "query",
            "description": "Narrows to one presentation policy.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List presentation transactions. Metadata only — never a result",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/presentations/{presentationId}": {
      "get": {
        "operationId": "PresentationController_get",
        "parameters": [
          {
            "name": "presentationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a presentation transaction and its result",
        "tags": [
          "Presentations"
        ]
      }
    },
    "/v1/presentations/{presentationId}/cancel": {
      "post": {
        "operationId": "PresentationController_cancel",
        "parameters": [
          {
            "name": "presentationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel a presentation transaction",
        "tags": [
          "Presentations"
        ]
      }
    },
...
  },
  "info": {
    "title": "European Digital Trust Platform — Verification as a Service",
    "description": "V0 business API. Customers work with presentation policies and transactions; protocol details (DCQL, OpenID4VP, engine sessions) are not part of this contract. No ARF or Technical Specification conformance and no production readiness is claimed.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tenant API key"
      }
    },
...
  }
}